Set up typescript and fix initial types and errors for the change
