+++ package.json
@@ -15,8 +15,8 @@
 		"react-router-dom": "^6.21.2"
 	},
 	"devDependencies": {
-		"@types/react": "^18.2.43",
-		"@types/react-dom": "^18.2.17",
+		"@types/react": "^18.2.48",
+		"@types/react-dom": "^18.2.18",
 		"@vitejs/plugin-react": "^4.2.1",
 		"autoprefixer": "^10.4.16",
 		"daisyui": "^4.6.0",
@@ -28,6 +28,7 @@
 		"stylelint": "^16.1.0",
 		"stylelint-config-standard": "^36.0.0",
 		"tailwindcss": "^3.4.1",
+		"typescript": "^5.3.3",
 		"vite": "^5.0.8"
 	}
 }
