+++ src/App.tsx
@@ -3,11 +3,11 @@ import { RouterProvider } from 'react-router-dom';
 import { router } from './config/router/router';
 
 function App() {
-	return (
-		<>
-			<RouterProvider router={router}></RouterProvider>
-		</>
-	);
+  return (
+    <>
+      <RouterProvider router={router}></RouterProvider>
+    </>
+  );
 }
 
 export default App;
