+++ src/components/CompanionIcon/CompanionIcon.tsx
