+++ src/main.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
+import App from './App'
 import './index.css'
 import './styles/cat-hero-styles.css';
 
-ReactDOM.createRoot(document.getElementById('root')).render(
+const rootElement = document.getElementById('root');
+ReactDOM.createRoot(rootElement || document.createElement('div')).render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
