+++ src/types/ICompanion.ts
@@ -1,0 +1,23 @@
+export interface IBaseCompanion {
+	name: string;
+	rarity: string;
+	image: string;
+	manaCost: number;
+	baseAttack: number;
+	baseASPD: number;
+	attackPerLevel: number;
+	ASPDPerLevel: number;
+	specialEffect: {
+		bonusId: number;
+		description: string;
+		tiers: {
+			level: number;
+			requiredLevel: number;
+			[key: string]: number;
+		}[];
+	}[];
+	getDescription?: (level?: number) => string[];
+}
+export interface ICompanion extends IBaseCompanion {
+	getDescription: (level?: number) => string[];
+}
